Added Wolverines and Ferrets and Pawnkind Loadouts
Added CE compatibility to Logann/Wolvx/Wolverine People and Ferretmen.

Added Pawnkind loadouts to:
- Leeani
- Racc
- Xenn
- Frijjjjjjid
- Fennekin
- Ferrets
- Woverines

Made the Xenn Cryogun into the Burngun and reduced their rates of fire. This is until I can make the Frost Cannon have it's own ammotype that isnt shared with the Orassans.

Not touched in this commit:
- Horse People (Lack of interest)
- Orassans (Lotta work required, will do later)